+++ test/just_audio_cache_test.dart
@@ -4,19 +4,13 @@ import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:just_audio/just_audio.dart';
 import 'package:just_audio_cache/just_audio_cache.dart';
-import 'package:just_audio_cache/src/parser.dart';
 import 'package:path_provider/path_provider.dart';
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
   group('audio_test', () {
-    test('test_cache_path', () async {
-      final url = 'https://baidu.com/hello/world.mp3';
-      String path = getUrlSuffix(url);
-      expect('helloworld.mp3', path);
-    });
-
-    /*test('download_test', () async {
+    test('download_test', () async {
+      final spMockCacheStrs = <String, String>{};
       final directory = await Directory.systemTemp.createTemp();
       // Mock out the MethodChannel for the path_provider plugin.
       const MethodChannel('plugins.flutter.io/path_provider')
@@ -26,30 +20,45 @@ void main() {
         if (methodCall.method == 'getApplicationDocumentsDirectory') {
           return directory.path;
         }
+        if (methodCall.method == 'getTemporaryDirectory') {
+          return directory.path;
+        }
         return null;
+      });
+      // Mock out the MethodChannel for the share_preference plugin.
+      const MethodChannel('plugins.flutter.io/shared_preferences')
+          .setMockMethodCallHandler((MethodCall methodCall) async {
+        // If you're getting the apps documents directory, return the path to the
+        // temp directory on the test environment instead.
+        if (methodCall.method == 'getAll') {
+          return spMockCacheStrs;
+        }
+        if (methodCall.method == 'setString') {
+          spMockCacheStrs[methodCall.arguments] = methodCall.arguments;
+          return true;
+        }
+        if (methodCall.method == 'getString') {
+          return spMockCacheStrs[methodCall.arguments];
+        }
       });
 
       final dir = await getApplicationDocumentsDirectory();
       final player = AudioPlayer();
-      await player.dynamicSet(
-          url: 'https://www.mboxdrive.com/m54.mp3',
-          path: dir.path,
-          excludeCallback: (_) => true);
-
-      final player2 = AudioPlayer();
-      await player2.cacheFile(url: 'https://www.mboxdrive.com/m54.mp3');
-
-      final player3 = AudioPlayer();
-      final url = 'https://www.mboxdrive.com/m54.mp3';
-      final existed = await player3.existedInLocal(url: url);
-      expect(true, existed);
-      if (await player3.existedInLocal(url: url)) {
-        expect(player3.getCachedPath(url: url),
-            dir.path + '/audio_cache/' + 'm54.mp3');
+      final sources = [
+        'https://www.mboxdrive.com/m54.mp3',
+        'https://www.mboxdrive.com/%c3%86STRAL%20-%20half%20light.mp3',
+        'https://www.mboxdrive.com/%ca%8f%e1%b4%80%20%ca%99%e1%b4%9c%e1%b4%85%e1%b4%9c.mp3'
+      ];
+      await player.dynamicSetAll(sources);
+      // await player.dynamicSet(
+      //     url: 'https://www.mboxdrive.com/m54.mp3',
+      //     path: dir.path,
+      //     excludeCallback: (_) => true);
+      for (final url in sources) {
+        expect(await player.existedInLocal(url: url), true);
       }
-    });*/
 
-    /*test('download_and_cache_file', () async {
+      /*test('download_and_cache_file', () async {
       final player = AudioPlayer();
       player.cacheFile(url: 'https://www.mboxdrive.com/m54.mp3');
     });
@@ -63,5 +72,6 @@ void main() {
             dir.path + '/audio_cache/' + 'm54.mp3');
       }
     });*/
+    });
   });
 }