+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-## 0.0.1
+## 0.1.0
 
-* TODO: Describe initial release.
+Official version
+
+### 0.1.1
+Add [dynamicSetAll] method
+* Store all the playlist file async with just 1 method